intermediate version of the revised version of the icp module, but with a non-termination bug

+++ src/lib/modules/ICPModule/ContractionCandidate.h
@@ -276,7 +276,7 @@ namespace smtrat
                 }
             }
 
-            unsigned activity()
+            size_t activity()
             {
                 return mOrigin.size();
             }
